Feedback changes - moved code to transforms.js
