+++ packages/block-library/src/post-title/edit.js
@@ -19,9 +19,7 @@ import { ToggleControl, TextControl, PanelBody } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { createBlock, getDefaultBlockName } from '@wordpress/blocks';
 import { useEntityProp, store as coreStore } from '@wordpress/core-data';
-import { useSelect, select } from '@wordpress/data';
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { store as editorStore } from '@wordpress/editor';
+import { useSelect } from '@wordpress/data';
 export default function PostTitleEdit( {
 	attributes: { level, levelOptions, textAlign, isLink, rel, linkTarget },
 	setAttributes,
@@ -182,43 +180,3 @@ export default function PostTitleEdit( {
 		</>
 	);
 }
-export const transforms = {
-	from: [
-		{
-			type: 'block',
-			blocks: [ 'core/heading', 'core/paragraph' ],
-			transform: ( attributes ) =>
-				createBlock( 'core/post-title', {
-					textAlign: attributes.align || undefined, // Map alignment if available
-				} ),
-		},
-	],
-	to: [
-		{
-			type: 'block',
-			blocks: [ 'core/heading' ],
-			transform: ( { textAlign } ) => {
-				const postTitle =
-					select( editorStore ).getEditedPostAttribute( 'title' ) ||
-					__( 'Add a title…' );
-				return createBlock( 'core/heading', {
-					content: postTitle,
-					align: textAlign,
-				} );
-			},
-		},
-		{
-			type: 'block',
-			blocks: [ 'core/paragraph' ],
-			transform: ( { textAlign } ) => {
-				const postTitle =
-					select( editorStore ).getEditedPostAttribute( 'title' ) ||
-					__( 'Add a title…' );
-				return createBlock( 'core/paragraph', {
-					content: postTitle,
-					align: textAlign,
-				} );
-			},
-		},
-	],
-};
